requirements.txt
Click==7.0
Flask==1.1.1
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
numpy==1.18.1
Pillow==7.0.0
six==1.13.0
torch==1.5.0.dev20200109+cpu
torchvision==0.5.0.dev20200109+cpu
Werkzeug==0.16.0
